Refactor Plain Output Writer
