+++ src/main/java/io/aiven/kafka/connect/common/grouper/RecordGrouper.java
@@ -27,6 +27,7 @@ import org.apache.kafka.connect.sink.SinkRecord;
 public interface RecordGrouper {
     /**
      * Associate the record with the appropriate file.
+     * @param record - record to group
      */
     void put(SinkRecord record);
 
@@ -37,6 +38,7 @@ public interface RecordGrouper {
 
     /**
      * Get all records associated with files, grouped by the file name.
+     * @return map of records assotiated with files
      */
     Map<String, List<SinkRecord>> records();
