+++ src/main/java/io/aiven/kafka/connect/common/output/plainwriter/AbstractValuePlainWriter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.aiven.kafka.connect.common.output;
+package io.aiven.kafka.connect.common.output.plainwriter;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -24,7 +24,7 @@ import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.errors.DataException;
 import org.apache.kafka.connect.sink.SinkRecord;
 
-public abstract class AbstractValueWriter implements OutputFieldWriter {
+public abstract class AbstractValuePlainWriter implements OutputFieldPlainWriter {
     /**
      * Takes the {@link SinkRecord}'s value as a byte array.
      *
