+++ src/main/java/io/aiven/kafka/connect/common/output/plainwriter/ValuePlainWriter.java
@@ -14,13 +14,11 @@
  * limitations under the License.
  */
 
-package io.aiven.kafka.connect.common.output;
+package io.aiven.kafka.connect.common.output.plainwriter;
 
-import java.util.Base64;
-
-public class Base64ValueWriter extends AbstractValueWriter {
+public class ValuePlainWriter extends AbstractValuePlainWriter {
     @Override
     protected byte[] getOutputBytes(final byte[] value) {
-        return Base64.getEncoder().encode(value);
+        return value;
     }
 }
